requirement.txt
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
SQLAlchemy==1.4.41
Werkzeug==2.2.2
zipp==3.8.1
datetime
uuid
PyJWT
Flask-Restful
Flask-Httpauth